+++ src/main/java/com/enn3developer/gtnhvoice/client/source/VoiceSourceManager.java
@@ -77,7 +77,7 @@ public final class VoiceSourceManager {
         VoiceSource source = sources.computeIfAbsent(packet.getSourceId(), uuid -> createSource(uuid, distance));
         if (source == null) return; // creation failed, already logged
 
-        source.handleAudio(packet.getData(), packet.getX(), packet.getY(), packet.getZ());
+        source.handleAudio(packet.getSequenceNumber(), packet.getData(), packet.getX(), packet.getY(), packet.getZ());
     }
 
     public void onSourceEnd(@NotNull SourceEndPacket packet) {